shared/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blender-shared"
version = "0.1.0"
description = "BlenderAgentBot shared protocol, config, and types"
requires-python = ">=3.10"
dependencies = [
    "pydantic",
    "pyyaml",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.hatch.build.targets.wheel]
packages = ["."]
